part3/requirements.txt
Flask
Flask-RESTX
Flask-Bcrypt
Flask-JWT-Extended
Flask-SQLAlchemy

